+++ src/index.ts
@@ -1,5 +1,4 @@
 export * from "./types";
-export * from "./ports";
 export { Connector } from "./Connector";